requirements.txt
flask==2.2.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.28.2
werkzeug==2.2.3
supabase==1.0.3
